Restructure middleware into a module
